Спринт 10 2 вариант 24.08.22

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -56,7 +56,7 @@ public class FilmController {
     }
 
     @GetMapping("films/{id}")
-    public Film getFilm(@PathVariable int id) {
+    public Film getFilm(@PathVariable long id) {
         log.info("Получен GET запрос на получение фильма");
         return filmService.getFilm(id);
     }
